+++ packages/components/pie-breadcrumb/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @justeattakeaway/pie-breadcrumb
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`4417c95`](https://github.com/justeattakeaway/pie/commit/4417c957e83f3a94018f1c7ea29d01d34ee70ebe)]:
+  - @justeattakeaway/pie-link@1.4.0
+
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [[`d7e5d5d`](https://github.com/justeattakeaway/pie/commit/d7e5d5d211be24a29cff33b077509d5cf4b6289b), [`d7e5d5d`](https://github.com/justeattakeaway/pie/commit/d7e5d5d211be24a29cff33b077509d5cf4b6289b)]:
+  - @justeattakeaway/pie-webc-core@16.1.0
+  - @justeattakeaway/pie-link@1.3.41
+  - @justeattakeaway/pie-icons-webc@1.27.1
+
 ## 1.0.5
 
 ### Patch Changes